requirements.txt
Django==1.11.18
django-webpack-loader==0.6.0
djangorestframework==3.6.3
pytz==2018.9
